Extract Tweet model converter
